[entropy_src/dv] Predict value of repcnt watermarks
Prior to this commit, Entropy Source's scoreboard did not check the
value read from the `repcnt_hi_watermarks` (i.e., bit repetition count
watermarks) and `repcnts_hi_watermarks` (i.e., symbol repetition count
watermarks) registers.

This commit models the internal repetition counters as well as their
propagation to the watermark register fields (which are separate for
FIPS and bypass mode) in the scoreboard and enables the check of the
watermark registers against the modeled value.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>