[manuf] introduce crt0 to setup stack, global pointer and clear BSS
Instead of relying on the opentitanlib JTAG loader to setup those
registers, create a small crt library that is linked into sram programs
and relies on the linker script to compute the right values.
This commit also create a new linker script since manufacturing should
not rely on the example linker script and will have its own requirements.
Since we are introducing a CRT, change the linker script to create a BSS
section that is not loaded and is cleared by the CRT. This saves about 25%
in the size of SRAM program which is also a save in the loading time over
JTAG.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>