[spi_host,dv] Extend sw_reset test to trigger mid-transaction
Refactor the test to spawn an additional thread which waits for a randomized
numbers of SCK edges before triggering a write that sets CONTROL.SW_RESET to 1.
If the randomization means we don't trigger the sw_reset during the transaciton,
then trigger it immediately after (when the DUT is idle).
We should gather coverage over sw_reset being triggered both in and out of an
active transaction.

This change required some previous changes to the scoreboard and base sequences
to allow a second thread to access the bus mid-transaction, and update the
checking to handle the on-the-fly reset behaviour.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>