[ci] Enable HyperDebug tests to run in CI
HyperDebug hardware is required for running these tests, so they were disabled
for CI runs. CI now has HyperDebug boards installed for all FPGAs, so these
should be capable of running there.

Applying a new `hyperdebug` tag so these can be skipped when working with
hardware that doesn't have HyperDebug connected (or when it breaks).

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>