[opentitantool] fix bugs in `image update` command
This fixes bugs in the `image update` command that was not updating the
manifest extentions if a SPX+ public key was passed on the cmd line, but
only a private key. If a public key was passed we still want to add the
extensions so we can sign with RSA (the signature of which includes the
manifest extentions), but allow us to attach the SPX signature later
when signing offline.

Signed-off-by: Tim Trippel <ttrippel@google.com>