[entropy_src/dv] Remove TODO on sigmas for bypass and FIPS modes
We don't need sigmas to be different between FIPS and bypass mode, as we
already distinguish the window size between FIPS and bypass mode.  So in
the current DV code, we derive the thresholds for FIPS and bypass mode
based on the same sigma but different window sizes, meaning the
threshold will differ between FIPS and bypass mode as long as the window
size is different.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>