[opentitanlib,opentitantools,test] migrate `structopt` to `clap`
`structopt` is in maintenance mode and should be migrated to `clap`
instead.

This is a fairly mechanical translation by replacing
`#[derive(StructOpt)]` with `#[derive(Args)]` on structs, and
`#[derive(Subcomamnd)]` on enums.

A few places using `PathBuf` are tweaked to use `PathBuf::from_str` instead
of the built-in value parser in clap, because clap's parser for
`PathBuf` rejects empty strings while we do need to accept that.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>