[edn/dv] Fix state sampling loop in `edn_disable_vseq`
When `edn_disable_vseq` should wait for a specific state of EDN's main
FSM, it has to sample the FSM state in every clock cycle.  To prevent
race conditions with events that change the FSM state, this ideally is
implemented with a `clocking` block, which we don't have in this case.
Alternatively, the DV env can sample signals of the DUT on the inactive
clock edge.

This commit changes the sampling of the main FSM state in
`edn_disable_vseq` to the negative clock edge.  It additionally changes
the loop from a `while` construct where the first value gets compared
before it gets sampled to a `forever` construct that does not have this
problem.  This commit also changes the wait statement in the other
branch of the surrounding `if`-`else` statement, so that both branches
are in-phase.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>