[edn/dv] Let `edn_err_vseq` randomly wait for short-lived FSM states
EDN's Main FSM and Ack FSM only stay for a single cycle in certain
states.  Consequentially, these states are very seldomly hit by a purely
random delay.

This commit changes `edn_err_vseq` so that it awaits a random state of
Main FSM or Ack FSM (with a probability of 90%) before it inserts an
error that triggers an escalation that affects those FSMs.  This
substantially increases the coverage of the transition from each of
those states to the Error state without needing more reseeds.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>