[release] Increase timeout of CI job
Even with a time limit of 5h (increased in
847f85160bcf396b9a5858cfcbc000a49ce380fa / PR #18952), the *Release* job
still times out.  In the latest log, a Verilator build runs before the
bitstream build and takes ca. 40min to complete.

This commit increases the timeout of the *Release* job to 7h to allow
for that Verilator build to complete in addition to two CW310 bitstream
builds and bitstream splicing.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>