[i2c,dv] Put host FIFO full TODO in ICEBOX
I am putting this in the icebox because the current test sends a random number of transactions with a random number of bytes and checks that the FIFO level, empty and full signals are exercised the right way.
The FIFO full level is hit when number of transactions times number of bytes is greater than 64, the FIFO depth.
Number of transactions varies from 10-20 and bytes varies from 1-32 so we will hit these interesting cases with reasonable likelihood without making a directed test case for it.

Issue URL: https://github.com/lowRISC/opentitan/issues/18979

Signed-off-by: Marno van der Maas <mvdmaas+git@lowrisc.org>