[rom] update sigverify mode behavior
We update the ROM to always read the OTP setting for what sigverify mode
to use (SW vs OTBN), regardless of the LC state (previously, the mode
setting was only read from OTP when the LC state was NOT
TEST_UNLOCKED*).

Additionally, we default the sigverify mode to SW, if the
`CREATOR_SW_CFG_SIGVERIFY_RSA_MOD_EXP_IBEX_EN` field OTP is
uninitialized (i.e., all 0s).

We make these changes because we want to run GLS tests in the
TEST_UNLOCKED* LC state using the real ROM, and not enabling the use of
OTBN signature verification causes GLS runtimes to explode.

Signed-off-by: Tim Trippel <ttrippel@google.com>