[device] Introduce mecanism to override status report list size.
We need a way to override the size for selected tests. This is
achieved by using some weak symbols and a new OTTF_* macro.
At the same, this commits reworks the overflow handling of the
list: previously it would just drop new entries when the list
is full. It makes more sense to keep the last N entries since
those are the most relevant to trace back an error. Update the
report unittest since now the printing order is reversed.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>