[otbn,dv] Remove stale TODO comment from stats.py
This is in "stats.py", which is only a helpful addition that made it
easier to get code statistics before we had RTL. We've still got the
stats code, but we're not really using it, I don't think. But drop the
TODO: we're not going to add a start address here.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>