[otbn,dv] Simplify wiping in gpr.py
The wipe() function wiped every register except x1. Presumably, that's
because x1 works differently (because it's the call stack) and I was
separating it out in my head when I wrote this.

Move the logic that wipes the call stack into gprs.wipe() (which is
only called from one place, so the code should behave the same).

Also remove a TODO about a possible check that we could add to make
sure the call stack is initially empty. This might be a good idea, but
a GitHub issue is probably a better place to track it.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>