[otbn,dv] Remove stale TODO in otbn_model_monitor.sv
I think I added comment because I imagined that we could change things
so that we only collected instruction logs when coverage was enabled.
This is a bit silly, because we want to know what instructions we're
running in order to compare them with the RTL, no matter whether or
not (code) coverage is enabled. Drop the comment.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>