[i2c/dv] Adjust host_perf constraints to reflect IP requirements
The IP requires a minimum thigh and tlow of 3 each. Adjust the
constraints to enable satisfying this requirement.

Target modes based on the input clock period, since it's impossible to
hit all specs for faster modes at lower frequencies. Float the SCL
period / frequency to allow thigh and tlow to adjust to the required
minimums, then calculate the coerced value and use it for comparisons.

Add accounting for how the IP currently tracks time and a TODO to fix
this later.

Signed-off-by: Alexander Williams <awill@opentitan.org>