[bazel] Add an aspect to check for colliding module IDs
We want to be able to run `opentitantool status lint` on every ELF
file that depends on //sw/device/lib/base:status. The approach here
is to use a bazel aspect to run an action for every ELF file. One
difficulty is that we have no easy way to specifying the list of
targets that we want to run the aspect on, and we want to avoid
modifying existing rules or list all targets by hand. A compromise
is to use a shell script to query the list of targets and then
manually specify an aspect on bazel's CLI when building them.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>