[rom_ctrl/dv] Increase TL access timeout
TL accesses are blocked while the ROM check is in progress.
The current timeout is large enough in most cases, but depending
on the randomization, there can be cases where the ROM check takes
longer than that default timeout.

This patch increases the timeout x4 to be on the safe side.

Signed-off-by: Michael Schaffner <msf@opentitan.org>