[manuf] Read back and verify device that performs the CRC check
At the moment we rely on the device code to verify the CRC of the
entire program at runtime. This catches most errors but in the unlikely
case where the corruption happens in the CRC code itself, the test could
pass even though the program is corrupted. This commit adds a feature to
the ELF loader to read back the data from the section that contains the
entry point of the program. The linker script is modified accordingly so
that a tiny section is created that contains the entry point and the only
the code to verify the CRC. Since this section is tiny, reading it back
has a negligible impact on loading time.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>