[manuf] Use non-page aligned sections for SRAM programs
At the moment, the ELF loader in opentitanlib has to load section
by sections instead of using segments. This is because by default
the first segment is page aligned and contain the ELF headers. This
is problematic because it can overwrite some important SRAM content
like static critical data. By using this option, we can prevent this
behaviour.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>