[otbn,dv] Don't apply loop warps in Verilator when no longer running
These "warps" are to bodge the internal loop counters in the RTL so
that they match the warping in the model.

The translation that they do doesn't work any more once the FSM in the
RTL has stopped, because the value of prefetch_loop_iterations_o
becomes 32'ffffffff.

Fortunately, we don't really care: we know we're stopping anyway.
Avoid bodging the loop counts in the RTL in that case: they'll never
have any more effect anyway.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>