[test] Fix activity-checking part of the power virus test
This part of the tests checks if all of the IPs are active during
the max-power-epoch.

This commit changes the activity indicator of the following blocks
to improve the stability of the test:

- Change the activity-indicator of CSRNG to the "acmd_q" signal, which
shows the last executed CSRNG operation. The chip_sw_power_virus_vseq.sv
file checks if the last CSRNG command is the reseed (0x2) or generate(0x3).

- Change the activity-indicator of AES to the round counter, which should
be greater than 0 in case the AES is running.

Tested on the local EDA farm instance.

Signed-off-by: Bilgiday Yuce <bilgiday@opentitan.org>