[dv, key_sideload_agent] Fix DV environment hanging issue under reset
Previously, the `get_and_drive()` task was consuming zero time when the
reset was active. This resulted in the key sideload interface endlessly
updating the sideload key and the whole DV environment to completely
hang. The root cause of this issue was the custom `wait_clks()`
implementation of the agent which 1) immediately returned when the reset
was active, and 2) was itself called in a forever loop consuming zero
time.

To fix this, this commit makes sure every iteration of forever loop
inside `get_and_drive()` takes at least one clock cycle. In line with
this change, the constraint for the response delay is adjusted to allow
for a minimum response delay of zero clock cycles (previously one). The
functions for waiting for clocks cycles and/or resets are copied over
from the clock and reset interface (`clk_rst_if.sv`) used in other
places of the project.

This is related to lowRISC/OpenTitan#19025.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>