[doc] Fix multireg documentation generation
The MultiRegister.reg attribute does not reflect the actual set of
registers and fields generated for a multireg description in hjson.
Change the documentation generation back to the full expansion of all
individual registers, since there is no guarantee that the expanded
registers even have the same number of fields or reset values.

Signed-off-by: Alexander Williams <awill@opentitan.org>