[regtool] Correct register QE assignment
This patch fixes cases where read-only and writable fields are mixed
in a single register, whereby the hwext register QE would have been
tied to 0 previously due to the logical AND of all field QEs.

The patch ignores field QEs that are always 0 when calculating the
register QE, except in the case where all fields are RO.

Fixes #19127.

Signed-off-by: Michael Schaffner <msf@opentitan.org>