[otbn,dv] Tweak length of time to wait for reset
Without this change, the following test fails when CDC is enabled:

  ./util/dvsim/dvsim.py hw/ip/otbn/dv/uvm/otbn_sim_cfg.hjson \
    -i otbn_alu_bignum_mod_err --fixed-seed 123 \
    --waves shm --run-only

It turns out that the failure is because it takes too long to get URND
data from the EDN. Looking at the waves, the timeout arrived when we
were partway through the first of three EDN requests.

If we wanted to make the test more robust here, we'd have to wait a
fixed time for each EDN request to come up and then allow the test
environment an arbitrary time to respond. But I don't expect this to
change again, so let's do the simple, silly fix for now.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>