[otbn,dv] Factor out some UVM spin-wait logic
In this vseq, we look at flags in the design and then do something
once they become true. Factor out the call to DV_SPINWAIT that we used
each time, and add a DV_CHECK_FATAL to assert that the calls to
uvm_hdl_read succeed (because the path is indeed valid).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>