[python] fix package dep that is incompatible with nightly infra
The `lxml` package is not a direct dependency of the OT project,
rather it is a transitive dependency. Since it is a transitive
dependency, it was not pinned to a specific version. As a result, a
newer version was pulled down onto our CentOS-based infrastructure that
runs the nightly regressions. The newer version was incompatible with
CentOS7, and therefore wiped out our nightly regressions. This makes the
last known compatible version a direct dependency of the project now to
fix the broken nightly regressions.

Signed-off-by: Tim Trippel <ttrippel@google.com>