[otbn,dv] Wait longer for model status change in otbn_scoreboard.sv
This code runs when we've seen an alert from the RTL and we're now
waiting for a status change (to locked) in the model. Since the
reported alert report and the status change aren't strictly tied
together, we have to allow a bit of slack between them: the gap is the
time taken to do a secure wipe between things going wrong and the
status switching to "locked".

With CDC enabled, this takes somewhat longer than the 400 cycles we
waited before. Bumping up by a factor of 10 seems to sort it out.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>