[otbn,dv] Fail properly in otbn_rf_base_intg_err_vseq if RF not used
This vseq is supposed to work by waiting until it sees the specified
register file in use. Once the register file is being read, it
corrupts the register file's response data and the test expects to see
OTBN catch the error.

If the register file doesn't actually get used before we run out of
time, this change makes the test fail more understandably. The error
message that was visible in the code doesn't actually appear in the
logs: it gets generated as UVM_HIGH by default. With this change, the
error message will appear as a fatal error: much easier to understand!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>