[otbn,dv] Fix logic when waiting in start_running_otbn
The start_running_otbn task in otbn_base_vseq is supposed to start
OTBN running and the task should complete when OTBN is still busy. The
idea is that we might want to e.g. inject an error into OTBN while
it's running.

There was a bug, and this code actually returned when OTBN was either
busy running something or was doing its end-of-run secure wipe. This
is weaker than callers of the task (otbn_rf_base_intg_err_vseq, for
example) might expect.

I only saw this at all because that sequence sometimes failed. It
turns out that its "guess a sensible time to stop" logic was sometimes
unlucky and guessed a time when OTBN doing its final secure wipe. The
rest sequence then wanted to inject an error on the register file, but
that didn't work: the register file wasn't used again!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>