[otbn,dv] Fix missing connections when binding in otbn_loop_if
This causes a warning in the build log (from xmelab with Xcelium) and
causes these signals to have value 'z instead of 0 or 1. Once that has
happened, any test with a loop warp will fail (because it checks that
we're not pushing/popping the loop stack and the check doesn't pass
when the signal happens to be 'z).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>