[kmac, rtl] Keep advancing LFSR in StRandErr as during normal operation
Previously, the LFSR wasn't advanced anymore and as a result, the masked
Keccak core would continuously use the same entropy while finishing the
current operation. This is bad from an SCA hardening perspective.

This commit thus changes the design to advance the LFSR as in regular
operation. The only difference is that the auxiliary randomness isn't
updated anymore but this is much less critical.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>