[usbdev,dv] Rename clocks in TB to fix various base tests
Before this commit, we had two clocks in the testbench:

    usb_clk (48MHz)    : Wired to "clk_i"
    clk (faster)       : Wired to "clk_aon_i"

The problem was caused by the fact that "clk" was also used as the
main clk/rst interface. We've got various DV code that assumes this
main clock interface (accessible through cfg.clk_rst_vif) is the one
used for things like alert propagation.

In particular, we've got checks that make sure a fatal alert gets
re-asserted every 6 cycles. To count these cycles, we uses the "main
clock". Before this change, the "main clock" that we were tracking was
faster than the one that the prim_alert_sender was using, which meant
that the alert didn't get re-asserted quickly enough.

This commit renames things slightly so that we have an explicit
usb_clk and aon_clk, but the "main clock" of the device is now the
thing that got passed in as clk_i, so the alert checks described above
now work properly.

A separate problem with the testbench is that the usb clock is
currently configured to be slower than the AON clock, which is the
wrong way around. That relationship isn't changed by this commit, but
we'll fix it up to be more sensible in a followup.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>