[usbdev,dv] Fix speed of AON clock
Before this patch, the AON clock was faster than the 48MHz USB clock.
This is backwards: it's supposed to be much slower (and low power!).
Switch the two around.

The real speed of the AON clock is going to be even slower than is
configured here, but having the two be reasonably similar should give
the same test efficacy and make the tests quicker to run (and debug!).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>