[kmac, rtl] Tweak SVA checking that the Keccak core only uses valdi PRD
If a reseeding operation throughout the randomness was valid finishes
just before the Keccak core starts processing, it may use randomness
which wasn't declared as valid (the valid is always de-asserted upon
finishing a reseed). The same LFSR output is then used for updating
the auxiliary randomness registers and for re-masking. This isn't ideal
but it this happens very rarely and thus should be okay.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>