[dv] Add list of "base CSRs" to CIP scoreboard
Recent changes to the code have added some general CSRs to all
comportable IPs. It turns out that the testbenches for these IPs have
scoreboards that (by default) throw an error if they see an access for
a CSR that isn't on a specified list. Of course, things like "cip_id"
aren't on the list!

Fortunately, the code that generates the errors came from a
template (scoreboard.sv.tpl) so it's not too hard to make things a bit
more general. This commit changes this code so that it doesn't special
case CSRs if they are in a known list.

This isn't as general as it could be: there are some IPs which will
still fail at the moment (because their scoreboards are more different
from the template, so will take a bit more effort to get in line).
Also, there are several other more general CSRs (like ALERT_TEST) that
might be nice to handle in a uniform way.

But this should be a minimal-ish change that will avoid having to
rewrite DV code for quite so many IP blocks.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>