[fpga] Add support for Earl Grey on CW340
Add FPGA build support for the CW340 + CW341 with Earl Grey. Port the
SPI constraints from the ASIC to the CW340, and move the USB timing
constraints out of the physical constraints file.

Adjust build scripts for the CW340's primitives for tasks like MMI
generation.

Use CW310 software for now. The device-side software should be mostly
compatible, but this could change in the future.

Signed-off-by: Alexander Williams <awill@opentitan.org>